split v3 schema field into service field and basic schema field
